+++ plugins/lambda-plugin/src/hooks/info.ts
@@ -29,7 +29,7 @@ export default async (context: InfoContext) => {
     console.log(`    - AccountId: ${account?.id}`);
     console.log(`    - Region: ${region}`);
 
-    context.output.functionInfo = await getFunction(lambdaClient, functionName, true);
+    context.output.functionInfo = await getFunction(lambdaClient, functionName, alias.name, true);
     if (!context.output.functionInfo) {
         return;
     }